finished retroactive compensation refactor
