+++ packages/i18n/locales/bad/translation.json
@@ -1062,8 +1062,6 @@
     "nothingHereYet": "bad bad bad",
     "notificationsInInbox_one": "bad bad bad bad bad bad bad",
     "notificationsInInbox_other": "bad bad bad bad bad bad bad",
-    "numContributors_one": "bad bad",
-    "numContributors_other": "bad bad",
     "numNftsSelected_one": "bad bad bad",
     "numNftsSelected_other": "bad bad bad",
     "numNftsSelected_zero": "bad bad bad",
@@ -1073,6 +1071,8 @@
     "numProposals_other": "bad bad",
     "numSelected_one": "bad bad",
     "numSelected_other": "bad bad",
+    "numSubmissions_one": "bad bad",
+    "numSubmissions_other": "bad bad",
     "onlyMembers": "bad bad",
     "openMobileWalletAndAcceptConnection": "bad bad bad bad bad bad bad bad bad",
     "openProfileSidebarTooltip": "bad bad bad bad bad",
