+++ packages/i18n/locales/en/translation.json
@@ -71,7 +71,9 @@
     "addedRegisterAction": "Added register action",
     "apolloSafe": "Apollo Safe",
     "approve": "Approve",
+    "back": "Back",
     "begin": "Begin",
+    "beginRating": "Begin rating",
     "burn": "Burn",
     "cancel": "Cancel",
     "castYourVote": "Cast your vote",
@@ -132,7 +134,7 @@
     "disable": "Disable",
     "done": "Done",
     "downloadMembersCsv": "Download member data (.csv)",
-    "downloadSurveyCsv": "Download survey data (.csv)",
+    "downloadResultsCsv": "Download results (.csv)",
     "downloadVotesCsv": "Download voting data (.csv)",
     "duplicate": "Duplicate",
     "earlyExecute": "Early execute",
@@ -190,6 +192,7 @@
     "numNotifications_zero": "Notifications",
     "openInChainExplorer": "Open in chain explorer",
     "openInDestination": "Open in {{destination}}",
+    "openProposal": "Open proposal",
     "openSidebarToVote": "Open sidebar to vote",
     "pause": "Pause",
     "prepareUpdate": "Prepare update",
@@ -274,6 +277,7 @@
     "viewDaosPage": "View DAO's page",
     "viewOnDaoDao": "View on DAO DAO",
     "viewProposals": "View proposals",
+    "viewResults": "View results",
     "withDecodedModuleMessages": "with Decoded Module Messages",
     "withdrawAvailableBalance": "Withdraw available balance",
     "withdrawFiat": "Withdraw fiat",
@@ -387,6 +391,7 @@
     "cannotStakeMoreThanYouHave": "You can't stake or unstake more tokens than you have.",
     "cannotTxZeroTokens": "You can't stake, unstake, or claim zero tokens.",
     "cannotUseCreateNftCollectionOnStargaze": "You cannot create a new NFT Collection via this action on Stargaze. Use the Stargaze Studio app on the DAO's Apps page instead.",
+    "cantDownloadRetroactiveSurveyCsv": "You must have been a member of the DAO when the compensation cycle was created in order to download the results.",
     "chainNotConnected": "Chain not connected.",
     "checkInternetOrTryAgain": "Check your internet connection, refresh the page, or try again later.",
     "closeDateMustBeAfterOpenDate": "Close date must be after open date.",
@@ -423,6 +428,7 @@
     "failedToGetTokenInfo": "Failed to get token info. Please enter a valid {{tokenType}} contract address.",
     "failedToLoadIbcConnection": "Failed to load IBC connection.",
     "failedToLoadIpfsProposalSave": "Failed to load proposal save from IPFS.",
+    "failedToLoadMembershipRefreshPage": "Failed to load DAO membership status. Refresh the page and try again.",
     "failedToLoadWidget": "Failed to load widget.",
     "failedToRelayAcks": "Failed to relay acknowledgements from {{chain}}.",
     "failedToRelayAndGetAcks": "Failed to relay packets and get acknowledgements.",
@@ -980,7 +986,10 @@
     "closedAtDate": "Closed at {{date}}",
     "closesAtDate": "Closes at {{date}}",
     "communityPoolDepositDescription": "Deposit funds into the community pool.",
-    "compensationCycleClosedAwaitingCompletion": "Contributions and ratings have been submitted. Review the results and create a proposal to complete the cycle. Alternatively, you may link it to an existing proposal or complete it without creating a proposal.",
+    "compensationCycleAwaitingRating_member": "Contributions have been submitted, and now DAO members must rate them. Review the submissions and submit your ratings.",
+    "compensationCycleAwaitingRating_notMember": "Contributions have been submitted and are now being rated by DAO members.",
+    "compensationCycleClosedAwaitingCompletion_member": "Contributions and ratings have been submitted. Review the results and create a proposal to complete the cycle. Alternatively, you may link it to an existing proposal or complete it without creating a proposal.",
+    "compensationCycleClosedAwaitingCompletion_notMember": "Contributions and ratings have been submitted, and DAO members are now reviewing the results.",
     "completeRelayOrLoseFunds": "Make sure you have completed the relaying process, or you may lose the tokens you sent to the relayers to pay fees.",
     "completed": "Completed",
     "configureRebalancerDescription": "Configure the treasury rebalancer.",
@@ -1092,6 +1101,7 @@
     "groupAddress": "CW4 Group",
     "highestUsdValue": "Highest USD value",
     "historySinceDate": "History since {{date}}",
+    "howCompleteCycle": "How do you want to complete this cycle?",
     "ibcTransferPathTooltip": "There is only one IBC hop in this spend.",
     "ibcTransferPathTooltip_pfm": "The IBC hops in this spend utilize packet-forward-middleware to send tokens through several chains in one transaction.",
     "icaExecuteDescription": "Execute an action from an ICA account on another chain.",
@@ -1206,8 +1216,6 @@
     "nothingHereYet": "Nothing here yet.",
     "notificationsInInbox_one": "There is {{count}} notification in your inbox.",
     "notificationsInInbox_other": "There are {{count}} notification in your inbox.",
-    "numContributors_one": "{{count}} contributor",
-    "numContributors_other": "{{count}} contributors",
     "numNftsSelected_one": "{{count}} NFT selected",
     "numNftsSelected_other": "{{count}} NFTs selected",
     "numNftsSelected_zero": "No NFT selected",
@@ -1217,6 +1225,8 @@
     "numProposals_other": "{{count}} proposals",
     "numSelected_one": "{{count}} selected",
     "numSelected_other": "{{count}} selected",
+    "numSubmissions_one": "{{count}} submission",
+    "numSubmissions_other": "{{count}} submissions",
     "onlyMembers": "Only members",
     "openMobileWalletAndAcceptConnection": "<0>Open your mobile wallet</0> and accept the connection request.",
     "openProfileSidebarTooltip": "Click to open profile sidebar.",
@@ -1319,6 +1329,7 @@
     "resumeRebalancerDescription": "Resume rebalancing.",
     "resumeRebalancerExplanation": "This action resumes rebalancing.",
     "retroactiveCompensationDescription": "After each contribution cycle, contributors are assigned points based on their contributions. Payment is split based on the points they receive.",
+    "retroactiveDidNotSubmit": "You did not submit a contribution this cycle.",
     "reviewActionImportData": "Review the following actions to make sure they look right. If they do, click the Import button at the bottom to add them to the proposal.",
     "reviewYourProposal": "Review your proposal...",
     "reviewYourTransaction": "Review your transaction...",
@@ -1863,6 +1874,7 @@
     "push": "Push",
     "pushNotifications": "Push Notifications",
     "quorum": "Quorum",
+    "rate": "Rate",
     "ratioOfVotes": "Ratio of votes",
     "rawData": "Raw data",
     "rebalancer": "Rebalancer",
