+++ packages/stateful/widgets/widgets/RetroactiveCompensation/Renderer/components/stateful/pages/ViewSurvey/index.ts
@@ -1,0 +1,1 @@
+export * from './ViewSurvey'
