+++ packages/stateful/widgets/widgets/RetroactiveCompensation/Renderer/components/stateless/TabRenderer.stories.tsx
@@ -1,12 +1,8 @@
 import { ComponentMeta, ComponentStory } from '@storybook/react'
 
-import { IconButtonLink } from '@dao-dao/stateless'
 import { DaoPageWrapperDecorator } from '@dao-dao/storybook'
 
-import {
-  Default as NewStoryFormStory,
-  makeProps as makeNewSurveyFormProps,
-} from './NewSurveyForm.stories'
+import { ButtonLink } from '../../../../../../components'
 import { TabRenderer } from './TabRenderer'
 
 export default {
@@ -22,43 +18,6 @@ const Template: ComponentStory<typeof TabRenderer> = (args) => (
 
 export const Default = Template.bind({})
 Default.args = {
-  loadingActiveSurveys: { loading: false, data: [] },
-  loadingCompletedSurveys: {
-    loading: false,
-    data: [
-      {
-        id: 1,
-        name: 'October 2022 Contributor Drop',
-        contributionCount: 10,
-        contributionsOpenedAt: '2022-10-01T12:00:00.000Z',
-        proposalId: 'A1',
-        createdAtBlockHeight: 1,
-      },
-      {
-        id: 2,
-        name: 'November 2022 Contributor Drop',
-        contributionCount: 7,
-        contributionsOpenedAt: '2022-11-01T12:00:00.000Z',
-        proposalId: 'A2',
-        createdAtBlockHeight: 2,
-      },
-      {
-        id: 3,
-        name: 'December 2022 Contributor Drop',
-        contributionCount: 14,
-        contributionsOpenedAt: '2022-12-01T12:00:00.000Z',
-        proposalId: 'A3',
-        createdAtBlockHeight: 3,
-      },
-    ],
-  },
-  loadingMembershipDuringCompletedSurveys: {
-    loading: false,
-    errored: false,
-    data: ['0', '5', '10'],
-  },
   isMember: true,
-  NewSurveyForm: () => <NewStoryFormStory {...makeNewSurveyFormProps()} />,
-  downloadCompletedSurvey: async (survey) => alert('download ' + survey.name),
-  IconButtonLink,
+  ButtonLink,
 }
