+++ packages/stateful/widgets/widgets/RetroactiveCompensation/Renderer/components/stateless/pages/ViewSurvey/Submit.tsx
@@ -8,30 +8,51 @@ import { Entity, LoadingData, TransProps } from '@dao-dao/types'
 import { formatDateTimeTz } from '@dao-dao/utils'
 
 import {
-  ActiveSurveyStatus,
   ContributionFormData,
   SurveyStatus,
-} from '../../../types'
-import { ContributionFormInput } from '../ContributionFormInput'
+  SurveyWithMetadata,
+} from '../../../../types'
+import { extractContributionFormDataFromSurvey } from '../../../../utils'
+import { ContributionFormInput } from '../../ContributionFormInput'
 
 export type SubmitProps = {
+  /**
+   * Whether or not a wallet is connected.
+   */
   connected: boolean
-  survey: ActiveSurveyStatus
+  /**
+   * The active survey.
+   */
+  status: SurveyWithMetadata
+  /**
+   * Callback to submit form data.
+   */
   onSubmit: (data: ContributionFormData) => Promise<void>
+  /**
+   * Whether or not the form is being submitted.
+   */
   loading: boolean
+  /**
+   * The wallet entity.
+   */
   loadingEntity: LoadingData<Entity>
+  /**
+   * Wallet entity display.
+   */
   EntityDisplay: ComponentType
+  /**
+   * Connect wallet button.
+   */
   ConnectWallet: ComponentType
+  /**
+   * Stateful i18n translation component.
+   */
   Trans: ComponentType<TransProps>
 }
 
 export const Submit = ({
   connected,
-  survey: {
-    survey,
-    contribution: existingContribution,
-    contributionSelfRatings,
-  },
+  status,
   onSubmit,
   loading,
   loadingEntity,
@@ -39,37 +60,15 @@ export const Submit = ({
   ConnectWallet,
   Trans,
 }: SubmitProps) => {
+  const { survey } = status
+
   const { t } = useTranslation()
 
-  let defaultContribution = existingContribution || ''
-  // Pull images out of the contribution text.
-  const defaultImages = defaultContribution
-    ? defaultContribution
-        .split('\n\n')
-        .pop()
-        ?.split('\n')
-        .map((part) =>
-          part.startsWith('![') ? part.split('](')[1].slice(0, -1) : undefined
-        )
-        .flatMap((url) => (url ? { url } : []))
-    : []
-  // If images were found, remove them from the text.
-  if (defaultImages?.length) {
-    defaultContribution = defaultContribution
-      .split('\n\n')
-      .slice(0, -1)
-      .join('\n\n')
-  }
-
   const formMethods = useForm<ContributionFormData>({
-    defaultValues: {
-      contribution: defaultContribution,
-      images: defaultImages,
-      ratings: contributionSelfRatings || survey.attributes.map(() => null),
-    },
+    defaultValues: extractContributionFormDataFromSurvey(status),
   })
 
-  const contributed = !!existingContribution
+  const contributed = !!status.contribution
 
   const dateText =
     survey.status === SurveyStatus.Inactive
